+++ frontend/modules/scenarioBuilder/components/scenarioBuilderItem.jsx
@@ -35,8 +35,9 @@ const ScenarioBuilderItem = ({
   onActionClicked,
   onRequestAccessClicked
 }) => {
+
   const newLayerIndex = layerIndex + 1;
-  console.log(actionType);
+
   return (
     <div className={isDeleting ? 'opacity-20' : ''} style={{ scrollMarginTop: "50px" }} id={`scenario-builder-slide-${slide._id}`}>
       <div className="flex justify-center">
